add lab data display, fixed initial display, fixed some migration bugs

+++ db/migrate/20160405174939_create_cbcs.rb
@@ -2,7 +2,7 @@ class CreateCbcs < ActiveRecord::Migration
   def change
     create_table :cbcs do |t|
       t.integer :wbc
-      t.integer :hgb
+      t.float :hgb
       t.integer :hct
       t.integer :plt
       t.references :visit, index: true, foreign_key: true
