+++ db/migrate/20160405183251_create_bmps.rb
@@ -1,13 +1,13 @@
 class CreateBmps < ActiveRecord::Migration
   def change
     create_table :bmps do |t|
-      t.integer :na
-      t.integer :k
-      t.integer :cl
-      t.integer :bicarb
-      t.integer :bun
-      t.integer :cr
-      t.integer :glucose
+      t.float :na
+      t.float :k
+      t.float :cl
+      t.float :bicarb
+      t.float :bun
+      t.float :cr
+      t.float :glucose
       t.references :visit, index: true, foreign_key: true
       t.timestamps null: false
     end
